requirements.txt
flask>=2.0.0
tweepy>=4.0.0
pillow>=9.0.0
python-dotenv>=0.19.0
openai>=0.27.0
numpy>=1.21.0
torch>=1.9.0
transformers>=4.15.0
python-magic>=0.4.24
requests>=2.26.0
boto3>=1.26.0
flask-wtf>=1.1.0 